Change mapbox internal access approach.
Since the patch wasn't accepted for the small parent change anyway, we will
need to use our own mapbox build so might as well make a more proper change.
This exports a bunch of stuff we might possibly want in the future and attaches
a `mapboxDereq` in the worker global scope to use as a require to access.